BookedChair backend implementering

+++ src/main/java/stud/kea/dk/biografbackend/customer/repository/CustomerRepository.java
@@ -3,6 +3,6 @@ package stud.kea.dk.biografbackend.customer.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import stud.kea.dk.biografbackend.customer.model.CustomerModel;
 
-public interface CustomerRepository extends JpaRepository<CustomerModel, Long> {
+public interface CustomerRepository extends JpaRepository<CustomerModel, Integer> {
 
 }
